requirements.txt
docpack>=0.1.2,<1.0.0
PyGithub>=2.6.1,<3.0.0
